Refactored and added cleanup
