+++ inc/cbd_render.h
@@ -1,0 +1,13 @@
+#ifndef CBD_RENDER_H
+# define CBD_RENDER_H
+
+#define WHITE 		0xFFFFFFFF
+#define	OFF_WHITE	0xF0F0F0FF
+#define RED 		0xFF0000FF
+#define GREEN 		0x00FF00FF
+#define BLUE 		0x0000FFFF
+
+#define TILES 		(16 + 16)
+#define TILESIZE 	WIDTH / TILES
+
+#endif //CBD_RENDER_H
