+++ src/parser/init.c
@@ -8,7 +8,7 @@ t_map	*alloc_map(void)
 	t_map	*mapdata;
 
 	mapdata = malloc(sizeof(t_map ));
-	mapdata->tex_path 		= ft_calloc(sizeof(char *), TEX_SIZE);
+	mapdata->tex_path 		= ft_calloc(sizeof(char *), TEX_SIZE + 1);
 	mapdata->cbd_map 		= NULL;
 	mapdata->raw_data 		= NULL;
 	mapdata->cbd_tex		= NULL;
